lambda/content/requirements.txt
boto3==1.24.5
bs4==0.0.1
pytz==2022.1
requests==2.27.1
newsapi-python==0.2.6